requirements.txt
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-cogwheels==0.3
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-modelcluster==5.0.1
django-recaptcha==2.0.6
django-taggit==1.2.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
draftjs-exporter==2.1.7
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
l18n==2018.5
msgpack==0.6.2
Pillow==6.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.8
Willow==1.3
zope.interface==5.1.0
whitenoise==5.1.0
